chore(prompts) add MdEditor to common
